+++ uoftscrapers/scrapers/athletics/utsc.py
@@ -1,4 +1,4 @@
-from ...scraper import Scraper
+from ..utils import Scraper
 from bs4 import BeautifulSoup
 from datetime import datetime
 from collections import OrderedDict
@@ -6,26 +6,26 @@ import json
 import requests
 
 
-class UTMAthletics:
-    """A scraper for the UTM athletics schedule.
+class UTSCAthletics:
+    """A scraper for the UTSC athletics schedule.
 
-    The schedule is located at http://www.utm.utoronto.ca/athletics/schedule
+    The schedule is located at http://www.utsc.utoronto.ca/athletics
     """
 
-    host = 'http://www.utm.utoronto.ca/athletics/schedule/month/'
+    host = 'http://www.utsc.utoronto.ca/athletics/calendar-node-field-date-time/month/'
     s = requests.Session()
 
     @staticmethod
     def scrape(location='.', month=None):
         """Update the local JSON files for this scraper."""
-        month = month or UTMAthletics.get_month(month)
+        month = month or UTSCAthletics.get_month(month)
 
-        Scraper.logger.info('UTMAthletics initialized.')
+        Scraper.logger.info('UTSCAthletics initialized.')
         headers = {
-            'Referer': UTMAthletics.host
+            'Referer': UTSCAthletics.host
         }
-        html = UTMAthletics.s.get('%s%s' % (UTMAthletics.host, month),
-                                  headers=headers).text
+        html = UTSCAthletics.s.get('%s%s' % (UTSCAthletics.host, month),
+                                   headers=headers).text
         soup = BeautifulSoup(html, 'html.parser')
 
         athletics = OrderedDict()
@@ -34,25 +34,27 @@ class UTMAthletics:
             for td in tr.find_all('td'):
                 date = td.get('data-date')
 
-                if not UTMAthletics.date_in_month(date, month):
+                if not UTSCAthletics.date_in_month(date, month):
                     continue
 
                 events = []
                 for item in td.find(class_='inner').find_all(class_='item'):
+                    title = item.find(class_='views-field-title').text.strip()
 
-                    # event cancelled or athletic center closed
-                    if item.find(class_='cancelled-item') or item.find(class_='athletics-calendar-note'):
-                        continue
+                    location_ = item.find(class_='views-field-field-location]')
 
-                    title = item.find(class_='athletics-calendar-title').text
-                    location_ = item.find(class_='athletics-calendar-location').text
+                    if location_.text.strip() == '':
+                        location_ = list(location_.next_siblings)[1]
+
+                    location_ = location_.text.strip()
+
                     start = item.find(class_='date-display-start').get('content')
                     end = item.find(class_='date-display-end').get('content')
 
                     events.append(OrderedDict([
                         ('title', title),
                         ('location', location_),
-                        ('building_id', '332'),
+                        ('building_id', '208'),
                         ('start_time', start),
                         ('end_time', end)
                     ]))
@@ -62,14 +64,10 @@ class UTMAthletics:
                     ('events', events)
                 ])
 
-        if athletics:
-            Scraper.ensure_location(location)
-
         for date, doc in athletics.items():
-            with open('%s/%s.json' % (location, date), 'w+') as outfile:
-                json.dump(doc, outfile)
+            Scraper.save_json(doc, location, date)
 
-        Scraper.logger.info('UTMAthletics completed.')
+        Scraper.logger.info('UTSCAthletics completed.')
 
     @staticmethod
     def get_month(m):
